% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationDBs-data.R, R/annotationDBs.R
\docType{data}
\name{annotationDBs}
\alias{annotationDBs}
\title{annotationDBs}
\format{list of org.db information.}
\source{
none
}
\usage{
data(annotationDBs)

annotationDBs
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
annotationDBs is an R object that contains information about organism 
specific mapping databases in BioConductor. Used to map gene identifiers.

annotationDBs is an R object that contains information about organism 
specific mapping databases in BioConductor. Used to map gene identifiers.
}
\examples{
data(annotationDBs)
source("annotationDBs.R")
}
\references{
See individual BioConductor pages.

none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{annotation}
\keyword{database}
\keyword{datasets}
\keyword{entrez}
\keyword{org.db}
