% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildNetwork}
\alias{buildNetwork}
\title{buildNetwork}
\usage{
buildNetwork(hitpredict, mygenes, taxid = 10090)
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Given a HitPredict data.frame (PPIs) and a vector of entrez IDs, build a
protein-protein interaction graph.
}
\examples{
buildNetwork()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
