% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiled-iPSD-data.R
\docType{data}
\name{compiled_iPSD}
\alias{compiled_iPSD}
\title{compiled iPSD proteome}
\format{atomic vector of entrez IDs (integers)}
\source{
Uezu et al., 2013
}
\usage{
data(compiled_iPSD)
}
\description{
A proteome of inhibitory synapses compile from several studies.
}
\examples{
data(compiled_iPSD)
}
\references{
Uezu et al. (2013) Science 353(6304):1123-9.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/27609886}{PubMed})
}
\keyword{datasets}
