% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ePSD-data.R
\docType{data}
\name{ePSD}
\alias{ePSD}
\title{ePSD proteome}
\format{atomic vector (integers) - Entrez IDs}
\source{
none
}
\usage{
data(ePSD)
}
\description{
The excitatory synapse postsynaptic proteome (iPSD) identified
by Uezu et al., (2016) by PSD-95 iBioID.
}
\examples{
data(ePSD)
}
\references{
Uezu et al. (2013) Science 353(6304):1123-9.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/27609886}{PubMed})
}
\keyword{datasets}
