% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommunity.R
\name{getCommunity}
\alias{getCommunity}
\title{getCommunity}
\usage{
getCommunity(graph, seeds, k = 2)
}
\arguments{
\item{graph}{- igraph object}
}
\value{
none
}
\description{
Get community of proteins that interact with seed proteins.
}
\examples{
getCommunity(graph, seeds)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
