% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComplexes.R
\name{getComplexes}
\alias{getComplexes}
\title{getComplexes}
\usage{
getComplexes(download = getwd(), quiet = TRUE)
}
\arguments{
\item{dataset}{- (character) which dataset to be downloaded from HitPredict.}
}
\value{
data - (data.table) HitPredict dataset with genes mapped to entrez.
}
\description{
Download a CORUM database.
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
