% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBs.R
\name{getDBs}
\alias{getDBs}
\title{getDBs()}
\usage{
getDBs()
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Returns a list of annotation databases.
}
\examples{
annotationDBs <- getDBs()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
