% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHitPredict.R
\name{getHitPredict}
\alias{getHitPredict}
\title{getHitPredict}
\usage{
getHitPredict(organism = "HitPredict", downloads = getwd(),
  keepdata = FALSE)
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Function that facilitates downloading the HitPredict database.
}
\examples{
none
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
