% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHomoloGene.R
\name{getHomoloGene}
\alias{getHomoloGene}
\title{getHomologene}
\usage{
getHomoloGene(hitpredict, taxid, downloads = getwd(), keepdata = FALSE)
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Downloading the NCBI holoogy database. Mapping proteins to species of 
interest.
}
\examples{
none
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
