% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHomologs.R
\name{getHomologs}
\alias{getHomologs}
\title{getHomologs}
\usage{
getHomologs(entrez, taxid, verbose = FALSE)
}
\arguments{
\item{taxid}{(integer) - taxonomic identifier for species of interest}

\item{hitpredict}{(data.table) - HitPredict data.}
}
\value{
none
}
\description{
Downloading the NCBI holoogy database. Mapping proteins to species of
interest.
}
\examples{
getHomologs(entrez, taxid = 10090)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
