% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs.R
\name{getIDs}
\alias{getIDs}
\title{getIDs}
\usage{
getIDs(
  identifiers,
  from,
  to,
  species = NULL,
  taxid = NULL,
  quiet = TRUE,
  multiVals = "first",
  ...
)
}
\arguments{
\item{identifiers}{- input gene identifiers}

\item{from}{- input identifier type, one of (case insensitive):
`c("ACCNUM", "ALIAS", "ENSEMBL", "ENSEMBLPROT",
"ENSEMBLTRANS", "ENTREZID", "ENZYME", "EVIDENCE", "EVIDENCEALL", "GENENAME",
"GO", "GOALL", "IPI", "MGI", "ONTOLOGY", "ONTOLOGYALL", "PATH",
"PFAM", "PMID", "PROSITE", "REFSEQ", "SYMBOL", "UNIGENE", "UNIPROT")`}

\item{to}{- output identifier type, see `from`.}

\item{species}{- organism identifier for input genes.}
}
\value{
output gene identifiers
}
\description{
Mapping gene identifiers.
An easy-to-use and robust wrapper around AnnotationDbi's mapIds function.
}
\examples{
getIDs(mygenes, from = "symbol", to = "entrez", species = "mouse")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
