% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInteractionMethods.R
\name{getInteractionMethods}
\alias{getInteractionMethods}
\title{getInteractionMethods}
\usage{
getInteractionMethods(hitpredict, quiet = TRUE)
}
\arguments{
\item{hitpredict}{(data.table)}
}
\value{
none
}
\description{
Given a HitPredict object (data.table), annotate the PPI-detection methods.
Note: A method score >= 0.485 is considered to indicate high confidence
according to Villaveces et al.
}
\examples{
getInteractionMethods(hitpredict)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{annotation}
\keyword{database}
\keyword{entrez}
\keyword{org.db}
