% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getMethods}
\alias{getMethods}
\title{getMethods}
\usage{
getMethods(hitpredict)
}
\arguments{
\item{hitpredict}{(data.table)}
}
\value{
none
}
\description{
Given a HitPredict object (data.table), annotate the PPI-detection methods.
Note: A method score >= 0.485 is considered to indicate high confidence
according to Villaveces et al.
}
\examples{
getMethods(hitpredict)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{annotation}
\keyword{database}
\keyword{entrez}
\keyword{org.db}
