% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPIs.R
\name{getPPIs}
\alias{getPPIs}
\title{getPPIs}
\usage{
getPPIs(organism = "HitPredict", mygenes, taxid, methods = "all",
  cutoff = 0, downloads = getwd(), keepdata = FALSE,
  saveppis = FALSE)
}
\arguments{
\item{organism}{(character) organism to be downloaded from HitPredict.
One of c("HitPredict",...)}

\item{mygenes}{(character or vector of integers) a character specifying the
file containing Entrez IDs for your genes of interest. Should contain a column,
Entrez with your genes. Alternatively, can be a vector of integers, Entrez IDs.}
}
\value{
none
}
\description{
Wrapper around getHitPredict, getHomoloGene, and getMethods.
Provided a list of genes, get PPIs amongst these proteins.
}
\examples{
getPPIs()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
