% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfile.R
\name{getProfile}
\alias{getProfile}
\title{getProfile}
\usage{
getProfile(g, partition_type, nsteps = 1, ...)
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Partition the network with the Leiden algorithm.
}
\examples{
getProfile()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
