% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxid.R
\name{getTaxid}
\alias{getTaxid}
\title{getTaxid
taxonomic id given a species alias}
\usage{
getTaxid(species)
}
\value{
none
}
\description{
getTaxid
taxonomic id given a species alias
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
