% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsInteractome_data.R
\docType{data}
\name{hsInteractome}
\alias{hsInteractome}
\title{hsInteractome data}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{}{}
}
\usage{
data("hsInteractome")
}
\description{
hsInteractome data
}
\examples{
data("hsInteractome") # load HitPredict human PPIs
}
\references{
HitPredict
}
\keyword{datasets}
