% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installDBs.R
\name{installDBs}
\alias{installDBs}
\title{installDBs}
\usage{
installDBs()
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Installation of organism specific mapping databases from BioConductor.
}
\examples{
installDBs()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
