% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPSD.R
\docType{data}
\name{iPSD}
\alias{iPSD}
\title{iPSD proteome}
\format{atomic vector (integers) - Entrez IDs}
\source{
none
}
\usage{
data(iPSD)
}
\description{
The inhibitory synapse postsynaptic proteome (iPSD) identified
by Uezu et al., (2016).
}
\examples{
data(iPSD)
}
\references{
Uezu et al. (2013) Science 353(6304):1123-9.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/27609886}{PubMed})
}
\keyword{datasets}
