% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{mapIDs}
\alias{mapIDs}
\title{mapIDs}
\usage{
mapIDs(identifiers, from, to, species, ...)
}
\arguments{
\item{identifiers}{}

\item{from}{- One of c("ACCNUM", "ALIAS", "ENSEMBL", "ENSEMBLPROT",
"ENSEMBLTRANS", "ENTREZID", "ENZYME", "EVIDENCE", "EVIDENCEALL", "GENENAME",
"GO", "GOALL", "IPI", "MGI", "ONTOLOGY", "ONTOLOGYALL", "PATH",
"PFAM", "PMID", "PROSITE", "REFSEQ", "SYMBOL", "UNIGENE", "UNIPROT")}

\item{to}{- see from}

\item{species}{}
}
\value{
none
}
\description{
Mapping gene identifiers.
}
\examples{
mapIDs(mygenes, from = "symbols", to = "entrez", species = "mouse")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
