% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi_batch_query.R
\name{mgi_batch_query}
\alias{mgi_batch_query}
\title{mgi_batch_query()}
\usage{
mgi_batch_query(ids, from = "Uniprot", to = "Entrez")
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
map uniprot IDs to entrez gene IDs using the MGI batch query tool.
}
\references{
http://www.informatics.jax.org/downloads/reports/MRK_Sequence.rpt
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
