% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musInteractome_data.R
\docType{data}
\name{musInteractome}
\alias{musInteractome}
\title{musInteractome data}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{}{}
}
\usage{
data("musInteractome")
}
\description{
musInteractome data
}
\examples{
data("musInteractome") # load HitPredict (human) PPIs mapped to mouse entrez
}
\references{
HitPredict
}
\keyword{datasets}
