% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synaptosome-data.R
\docType{data}
\name{synaptosome}
\alias{synaptosome}
\title{Synaptosome proteome}
\format{atomic vector (integers) Entrez IDs}
\source{
none
}
\usage{
data(synaptosome)
}
\description{
A synaptosome proteome.
}
\examples{
data(synaptosome)
}
\references{
none
}
\keyword{datasets}
