% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp-data.R
\docType{data}
\name{wrp}
\alias{wrp}
\title{Wrp iBioID proteome}
\format{atomic vector (integers) - Entrez IDs}
\source{
none
}
\usage{
data(iPSD)
}
\description{
The proteome of developing excitatory synapses identified by Wrp iBioID by
Spence et al., (2019).
}
\examples{
data(wrp)
}
\references{
Spence et al., (2019) Nature Communications, 10(1):386.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/30674877}{PubMed})
}
\keyword{datasets}
