% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationDBs.R
\docType{data}
\name{annotationDBs}
\alias{annotationDBs}
\title{annotationDBs}
\format{An object of class \code{list} of length 18.}
\usage{
annotationDBs
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
annotationDBs is an R object that contains information about organism 
specific mapping databases in BioConductor. Used to map gene identifiers.
}
\examples{
source("annotationDBs.R")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{annotation}
\keyword{database}
\keyword{entrez}
\keyword{org.db}
