% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationDBs-data.R
\docType{data}
\name{annotationDBs}
\alias{annotationDBs}
\title{annotationDBs}
\format{list of org.db information.}
\source{
none
}
\usage{
data(annotationDBs)
}
\description{
annotationDBs is an R object that contains information about organism 
specific mapping databases in BioConductor. Used to map gene identifiers.
}
\examples{
data(annotationDBs)
}
\references{
See individual BioConductor pages.
}
\keyword{datasets}
