% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHitPredict.R
\name{getHitPredict}
\alias{getHitPredict}
\title{getHitPredict}
\usage{
getHitPredict(dataset = "all", quiet = TRUE)
}
\arguments{
\item{dataset}{- (character) which dataset to be downloaded from HitPredict.
One of HitPredict (all interactions), H_sapiens, M_musculus, R_norvegicus,
A_thaliana, C_elegans, C_jejuni, D_melanogaster, E_coli, H_pylori,
P_falciparum, S_cerevisiae, or S_pombe.}
}
\value{
data - (data.table) HitPredict dataset with genes mapped to entrez.
}
\description{
Download a HitPredict database.
}
\examples{
getHitPredict(dataset = "HitPredict")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
