% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHitPredict.R
\name{getHitPredict}
\alias{getHitPredict}
\title{getHitPredict}
\usage{
getHitPredict(organism = "HitPredict")
}
\arguments{
\item{organism}{- (character) which organism to be downloaded from HitPredict.}

\item{downloads}{- (character) optional directory for downloaded data.}

\item{keepdata}{- (logical) should downloaded data be kept or removed?}
}
\value{
none
}
\description{
Function that facilitates downloading the HitPredict database.
}
\examples{
getHitPredict(organism = "HitPredict")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
