% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHitPredict.R
\name{getHitPredict}
\alias{getHitPredict}
\title{getHitPredict}
\usage{
getHitPredict(organism = "HitPredict")
}
\arguments{
\item{organism}{- (character) which organism to be downloaded from HitPredict.
One of HitPredict (all interactions), H_sapiens, M_musculus, R_norvegicus,
A_thaliana, C_elegans, C_jejuni, D_melanogaster, E_coli, H_pylori, P_falciparum,
S_cerevisiae, or S_pombe.}
}
\value{
none
}
\description{
Function that facilitates downloading the HitPredict database.
}
\examples{
getHitPredict(organism = "HitPredict")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
