% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHomoloGene.R
\name{getHomoloGene}
\alias{getHomoloGene}
\title{getHomologene}
\usage{
getHomoloGene(hitpredict, taxid, quiet = TRUE)
}
\arguments{
\item{hitpredict}{(data.table) - HitPredict data.}

\item{taxid}{(integer) - taxonomic identifier for species of interest}
}
\value{
none
}
\description{
Downloading the NCBI holoogy database. Mapping proteins to species of
interest.
}
\examples{
getHomoloGene(hitpredict, taxid = 10090)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
