% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHomologs.R
\name{getHomologs}
\alias{getHomologs}
\title{getHomologs}
\usage{
getHomologs(entrez, species = NULL, taxid = NULL, quiet = TRUE)
}
\arguments{
\item{entrez}{(character) - Entrez IDs -- can be from any species}

\item{hitpredict}{(data.table) - HitPredict data.}
}
\value{
none
}
\description{
Downloading the NCBI holoogy database. Mapping Entrez IDs to homologs in
a given species of interest.
}
\examples{
getHomologs(entrez, species = "mouse")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
