% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getMethods}
\alias{getMethods}
\title{getMethods}
\usage{
getMethods(hitpredict, path2mi, methods = "all", cutoff = 0.485)
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Given a HitPredict object (data.table), annotate the PPI-detection methods,
and refine PPI list by method type if provided.
}
\examples{
getMethods(hitpredict)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{annotation}
\keyword{database}
\keyword{entrez}
\keyword{org.db}
