% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPIs.R
\name{getPPIs}
\alias{getPPIs}
\title{getPPIs}
\usage{
getPPIs(dataset, species)
}
\arguments{
\item{dataset}{(character) dataset to be downloaded from HitPredict.
One of c("HitPredict",...)}

\item{species}{(character) species alias for organism of interest.}
}
\value{
none
}
\description{
Wrapper around getHitPredict, getHomoloGene, and getInteractionMethods.
}
\examples{
getPPIs()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
