% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPIs.R
\name{getPPIs}
\alias{getPPIs}
\title{getPPIs}
\usage{
getPPIs(organism = "HitPredict", mygenes, taxid = 10090, prots,
  methods = "all", cutoff = 0, file = "", downloads = getwd(),
  keepdata = FALSE)
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Wrapper around getHitPredict, getHomoloGene, and getMethods.
Provided a list of genes, get PPIs amongst these proteins.
}
\examples{
getPPIs()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
