% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPIs.R
\name{getPPIs}
\alias{getPPIs}
\title{getPPIs}
\usage{
getPPIs(organism, taxid)
}
\arguments{
\item{organism}{(character) organism to be downloaded from HitPredict.
One of c("HitPredict",...)}

\item{taxid}{(integer) taxonomic identifier for organism of interest.}
}
\value{
none
}
\description{
Wrapper around getHitPredict, getHomoloGene, and getMethods.
}
\examples{
getPPIs()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
