% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfile.R
\name{getProfile}
\alias{getProfile}
\title{getProfile}
\usage{
getProfile(g, partition_type, nsteps = 1, ...)
}
\arguments{
\item{g}{- igraph object}

\item{partition_type}{- Leiden algorithm method to be used.}

\item{nsteps}{- number of resolutions to be analyzed. If 1 then, 
the native resolution (r=0) will be used.}
}
\value{
none
}
\description{
Partition the network with the Leiden algorithm.
}
\examples{
getProfile()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
