% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs.R, R/mapIDs.R
\name{mapIDs}
\alias{mapIDs}
\title{mapIDs}
\usage{
mapIDs(
  identifiers,
  from,
  to,
  species = NULL,
  taxid = NULL,
  multiVals = "first",
  ...
)

mapIDs(
  identifiers,
  from,
  to,
  species = NULL,
  taxid = NULL,
  multiVals = "first",
  ...
)
}
\arguments{
\item{identifiers}{}

\item{from}{- One of c("ACCNUM", "ALIAS", "ENSEMBL", "ENSEMBLPROT",
"ENSEMBLTRANS", "ENTREZID", "ENZYME", "EVIDENCE", "EVIDENCEALL", "GENENAME",
"GO", "GOALL", "IPI", "MGI", "ONTOLOGY", "ONTOLOGYALL", "PATH",
"PFAM", "PMID", "PROSITE", "REFSEQ", "SYMBOL", "UNIGENE", "UNIPROT")}

\item{to}{- see from}

\item{species}{}
}
\value{
none

none
}
\description{
Mapping gene identifiers.
An easier to use wrapper around AnnotationDbi::mapIds().

Mapping gene identifiers.
An easier to use wrapper around AnnotationDbi::mapIds().
}
\examples{
mapIds(mygenes, from = "symbol", to = "entrez", species = "mouse")
mapIds(mygenes, from = "symbol", to = "entrez", species = "mouse")
}
\references{
none

none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}

Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
