% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi_batch_query.R
\name{mgi_batch_query}
\alias{mgi_batch_query}
\title{mgi_batch_query()}
\usage{
mgi_batch_query(
  ids,
  from = "Uniprot",
  to = "Entrez",
  download = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
map uniprot IDs to entrez gene IDs using the MGI batch query tool.
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
