% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musInteractome.R
\docType{data}
\name{musInteractome}
\alias{musInteractome}
\title{Mouse Interactome compiled from HitPredict}
\format{data.table - hitpredict data}
\source{
none
}
\usage{
data(musInteractome)
}
\description{
An interactome of mouse proteins compile from the HitPredict database.
}
\examples{
data(musInteractome)
}
\references{
none
}
\keyword{datasets}
